refactor: 💡 참여자 관리 로직을 cusstom hook(useParticipants)로 분리
